+++ ConvNetCS/Network/ParamsAndGrads.cs
@@ -9,10 +9,10 @@ namespace ConvNetCS
     [Serializable]
     public class ParamsAndGrads
     {
-        public double[] Params { get; set; }
-        public double[] Grads { get; set; }
-        public double l1_decay_mul { get; set; }
-        public double l2_decay_mul { get; set; }
+        public float[] Params { get; set; }
+        public float[] Grads { get; set; }
+        public float l1_decay_mul { get; set; }
+        public float l2_decay_mul { get; set; }
 
     }
 }
